+++ lib/controllers/auth_controller.dart
@@ -13,6 +13,7 @@ import '../services/nostr_mail_service.dart';
 import '../utils/toast_helper.dart';
 import 'package:flutter/material.dart';
 import 'inbox_controller.dart';
+import 'scheduled_controller.dart';
 import 'settings_controller.dart';
 
 class AuthController extends GetxController {
@@ -86,6 +87,9 @@ class AuthController extends GetxController {
       await Get.find<InboxController>().activateForCurrentAccount(
         folder: MailFolder.inbox,
       );
+    }
+    if (Get.isRegistered<ScheduledController>()) {
+      await Get.delete<ScheduledController>();
     }
     isLoggedIn.value = true;
     loadUserMetadata();
@@ -212,6 +216,9 @@ class AuthController extends GetxController {
     try {
       if (Get.isRegistered<InboxController>()) {
         await Get.find<InboxController>().resetForAccountChange();
+      }
+      if (Get.isRegistered<ScheduledController>()) {
+        await Get.delete<ScheduledController>();
       }
       await _nostrMailService.logout();
       await ndkFlutter.saveAccountsState();